Garbage Collection and Code Cleanup

+++ code/game/objects/items/weapons/explosives.dm
@@ -18,10 +18,11 @@
 	..()
 
 /obj/item/weapon/grenade/plastic/Destroy()
-	qdel(nadeassembly)
-	nadeassembly = null
+	if(nadeassembly)
+		qdel(nadeassembly)
+		nadeassembly = null
 	target = null
-	..()
+	return ..()
 
 /obj/item/weapon/grenade/plastic/attackby(obj/item/I, mob/user, params)
 	if(!nadeassembly && istype(I, /obj/item/device/assembly_holder))
