+++ code/modules/clothing/suits/hood.dm
@@ -10,7 +10,9 @@
 	..()
 
 /obj/item/clothing/suit/hooded/Destroy()
-	qdel(hood)
+	if(hood)
+		qdel(hood)
+		hood = null
 	return ..()
 
 /obj/item/clothing/suit/hooded/proc/MakeHood()
